Fix failing test
The need to mock the test for exists() on the `layer_file_path` is legmitimate becuase the lyr_props object is created with the verify_on_creation=False option. Other tests should ensure cover the more normal senario that  verify_on_creation=True (and therefore the .lyr files and layer properties match).